requirements.txt
requests
pushbullet.py
beautifulsoup
python-firebase
